refactor(cli): extract workflow scaffold handler
